ollegro_backend/requirements.txt
django==4.2.1
djangorestframework==3.14.0
django-cors-headers==3.13.0
gunicorn==20.1.0
psycopg2-binary
pydantic
djangorestframework-simplejwt==5.2.2
django-filter==23.2
